pyproject.toml
[tool.poetry]
name = "chatgpt-backend"
version = "0.1.0"
description = ""
authors = ["Pablo <ppmoya08@gmail.com>"]
readme = "README.md"
packages = [{include = "chatgpt_backend"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.92.0"
uvicorn = "^0.20.0"
openai = "^0.27.0"
pytest = "^7.2.1"
python-dotenv = "^1.0.0"
pymongo = "^4.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
